bai1b7

+++ buoi5-bai1.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main() {
     int a=3, b=5, c=2;
-    int tong = a+b+c;
+    long tong = a+b+c;
     cout<< "Tong la "<< tong <<endl;
     cout<< "Gia tri trung binh la "<< float(tong)/3<<endl;
     return 0;
